[CPU] Break LLVMCPUVectorLowering pass to several small passes.
The pass is break into four passes:

- LLVMCPUVirtualVectorLowering
- LLVMCPUVectorTransferLowering
- LLVMCPUVectorTransposeLowering
- LLVMCPUVectorShapeCastLowering

The revision creates a vector lowering pipeline, and a TODO about
running canonicalizer is addressed in the revision.
