+++ compiler/src/iree/compiler/Codegen/LLVMCPU/BUILD.bazel
@@ -66,6 +66,10 @@ iree_compiler_cc_library(
         "LLVMCPUTileAndFuse.cpp",
         "LLVMCPUUnfuseFMAOps.cpp",
         "LLVMCPUVectorLowering.cpp",
+        "LLVMCPUVectorShapeCastLowering.cpp",
+        "LLVMCPUVectorTransferLowering.cpp",
+        "LLVMCPUVectorTransposeLowering.cpp",
+        "LLVMCPUVirtualVectorLowering.cpp",
         "Passes.cpp",
         "TargetMLTransformInfo.cpp",
         "Utils.cpp",
