+++ compiler/src/iree/compiler/Codegen/LLVMCPU/CMakeLists.txt
@@ -67,6 +67,10 @@ iree_cc_library(
     "LLVMCPUTileAndFuse.cpp"
     "LLVMCPUUnfuseFMAOps.cpp"
     "LLVMCPUVectorLowering.cpp"
+    "LLVMCPUVectorShapeCastLowering.cpp"
+    "LLVMCPUVectorTransferLowering.cpp"
+    "LLVMCPUVectorTransposeLowering.cpp"
+    "LLVMCPUVirtualVectorLowering.cpp"
     "Passes.cpp"
     "TargetMLTransformInfo.cpp"
     "Utils.cpp"
